Remove login template into its own template
