+++ app/components/signUp/signUp.less
@@ -6,7 +6,7 @@
   max-width: 811px;
 }
 
-.sign-up-content {
+.sign-up-container {
   position: absolute;
   z-index: 1;
   color: @textColor;
@@ -59,7 +59,7 @@
   &:extend(.input-focus);
 }
 
-.sign-up-content:focus {
+.sign-up-container:focus {
   &:extend(.input-focus);
 }
